feat: blog page pulls from database
