+++ src/app/blog/page.tsx
@@ -2,6 +2,7 @@ import BlogPreview from "@/components/blogPreview";
 import connectDB from "@/database/db";
 import BlogModel from "@/database/blogSchema";
 import type { Blog } from "@/typings/blog";
+import { getBlogs } from "@/database/blogSchema";
 
 export default async function Blog() {
   const blogs = await getBlogs();
@@ -28,17 +29,4 @@ export default async function Blog() {
       ))}
     </div>
   );
-}
-
-async function getBlogs() {
-  await connectDB(); // function from db.ts before
-
-  try {
-    // query for all blogs and sort by date
-    const blogs = await BlogModel.find().sort({ date: -1 }).orFail();
-    // send a response as the blogs as the message
-    return blogs;
-  } catch (err) {
-    return null;
-  }
 }
